♻️ Refaktorer FullførKjørelisteFane med DataViewer og KjørelisteBrevmeny
- Ekstraher begrunnelse-logikk til KjørelisteBrevmeny-komponent
- Legg til useBegrunnelse-hook med Ressurs-mønster
- Bruk DataViewer for begrunnelseshenting
- Skjul brevmeny for ferdigstilte behandlinger
- Hent og vis eksisterende begrunnelse ved mount
- Legg til .catch på initial brevhenting
