+++ src/frontend/Sider/Behandling/Kjøreliste/FullførKjørelisteFane.tsx
@@ -1,11 +1,14 @@
 import React, { FC, useCallback, useEffect, useState } from 'react';
 
-import { Button, ExpansionCard, HStack, Textarea, VStack } from '@navikt/ds-react';
+import { Button, VStack } from '@navikt/ds-react';
 
 import styles from './FullførKjørelisteFane.module.css';
+import { KjørelisteBrevmeny } from './KjørelisteBrevmeny';
+import { useBegrunnelse } from './useBegrunnelse';
 import { useApp } from '../../../context/AppContext';
 import { useBehandling } from '../../../context/BehandlingContext';
 import { useSteg } from '../../../context/StegContext';
+import DataViewer from '../../../komponenter/DataViewer';
 import { Feilmelding } from '../../../komponenter/Feil/Feilmelding';
 import {
     erFeil,
@@ -15,6 +18,7 @@ import {
 } from '../../../komponenter/Feil/feilmeldingUtils';
 import { PdfVisning } from '../../../komponenter/PdfVisning';
 import {
+    byggRessursFeilet,
     byggTomRessurs,
     RessursFeilet,
     RessursStatus,
@@ -25,13 +29,11 @@ export const FullførKjørelisteFane: FC = () => {
     const { request } = useApp();
     const { behandling, behandlingErRedigerbar, hentBehandling } = useBehandling();
     const { erStegRedigerbart } = useSteg();
-    const [feilmelding, settFeilmelding] = useState<Feil>();
-    const [begrunnelseFeil, settBegrunnelseFeil] = useState<Feil>();
+    const { begrunnelseRessurs } = useBegrunnelse();
 
-    const [laster, settLaster] = useState<boolean>(false);
-    const [lasterBegrunnelse, settLasterBegrunnelse] = useState<boolean>(false);
-    const [begrunnelse, settBegrunnelse] = useState<string>('');
+    const [feilmelding, settFeilmelding] = useState<Feil>();
     const [harUlagredeEndringer, settHarUlagredeEndringer] = useState<boolean>(false);
+    const [laster, settLaster] = useState<boolean>(false);
 
     const [brevPdf, settBrevPdf] = useState(byggTomRessurs<string>());
 
@@ -41,7 +43,9 @@ export const FullførKjørelisteFane: FC = () => {
                 `/api/sak/kjorelistebrev/${behandling.id}`,
                 behandlingErRedigerbar ? 'POST' : 'GET',
                 behandlingErRedigerbar ? { begrunnelse: begrunnelseVerdi } : undefined
-            ).then(settBrevPdf);
+            )
+                .then(settBrevPdf)
+                .catch(() => settBrevPdf(byggRessursFeilet('Kunne ikke laste brev')));
         },
         [behandlingErRedigerbar, behandling.id, request]
     );
@@ -49,31 +53,6 @@ export const FullførKjørelisteFane: FC = () => {
     useEffect(() => {
         hentEllerGenererBrev(null);
     }, [hentEllerGenererBrev]);
-
-    const lagreBegrunnelse = () => {
-        if (lasterBegrunnelse) return;
-        settLasterBegrunnelse(true);
-        request<string, { begrunnelse: string | null }>(
-            `/api/sak/kjorelistebrev/${behandling.id}`,
-            'POST',
-            { begrunnelse: begrunnelse }
-        )
-            .then((res) => {
-                if (res.status === RessursStatus.SUKSESS) {
-                    settBrevPdf(res);
-                    settHarUlagredeEndringer(false);
-                    settBegrunnelseFeil(undefined);
-                } else {
-                    settBegrunnelseFeil(feiletRessursTilFeilmelding(res as RessursFeilet));
-                }
-            })
-            .catch((error) =>
-                erFeil(error)
-                    ? settBegrunnelseFeil(error)
-                    : settBegrunnelseFeil(lagFeilmelding('Ukjent feil oppstod'))
-            )
-            .finally(() => settLasterBegrunnelse(false));
-    };
 
     const fullfør = () => {
         if (harUlagredeEndringer) {
@@ -111,42 +90,15 @@ export const FullførKjørelisteFane: FC = () => {
         <div className={styles.toKolonner}>
             <VStack gap="space-16">
                 {erStegRedigerbart && (
-                    <div className={styles.background}>
-                        <ExpansionCard aria-label="Begrunnelse" size="small" defaultOpen>
-                            <ExpansionCard.Header>
-                                <HStack wrap={false} align="center" gap="space-8">
-                                    <ExpansionCard.Title size="small">
-                                        Begrunnelse
-                                    </ExpansionCard.Title>
-                                </HStack>
-                            </ExpansionCard.Header>
-                            <ExpansionCard.Content>
-                                <div className={styles.container}>
-                                    <Textarea
-                                        label="Begrunnelse for vedtaket"
-                                        hideLabel
-                                        value={begrunnelse}
-                                        onChange={(e) => {
-                                            settBegrunnelse(e.target.value);
-                                            settHarUlagredeEndringer(true);
-                                        }}
-                                        minRows={3}
-                                    />
-                                    <Feilmelding feil={begrunnelseFeil} />
-                                    <HStack>
-                                        <Button
-                                            variant="secondary"
-                                            loading={lasterBegrunnelse}
-                                            onClick={lagreBegrunnelse}
-                                            size="small"
-                                        >
-                                            Lagre begrunnelse
-                                        </Button>
-                                    </HStack>
-                                </div>
-                            </ExpansionCard.Content>
-                        </ExpansionCard>
-                    </div>
+                    <DataViewer type="begrunnelse" response={{ begrunnelseRessurs }}>
+                        {({ begrunnelseRessurs }) => (
+                            <KjørelisteBrevmeny
+                                initialBegrunnelse={begrunnelseRessurs.begrunnelse}
+                                settHarUlagredeEndringer={settHarUlagredeEndringer}
+                                settBrevPdf={settBrevPdf}
+                            />
+                        )}
+                    </DataViewer>
                 )}
 
                 {behandling.status === 'FERDIGSTILT' && (
